Wait on /updatetheme to prevent flakiness
Reloading after a theme change can race with the POST to /updatetheme,
which updates the session. The tests assume that the session has been
updated when they reload the page, and were therefore being flaky.

Waiting for the request ensures that the session has been updated
before the reload occurs, preventing the issue that causes flakiness.