Use waitForFormInteractive rather than .pipe(click)
Some early tests used .pipe(click) to hack around the fact the page
takes some time to load (mainly because the development server, is
slow as it compiles on the fly, but also because Cypress takes a
little while to load pages. I've since added a waitForFormInteractive
method that allows us to check that the page has loaded properly
rather than clicking repeatedly until we achieve the desired
result, or we time out.

This improves the tests in two main ways:
1. We actually verify that we achieve the desired result the
first time the user clicks the relevant button.
2. We reduce our dependency on cypress-pipe, to one usage, so we will
be able to remove cypress-pipe as a dependency soon (likely in the commit
that follows this).