Delete unused package.json for cypress
This is horribly outdated, and I don't think it ever really worked.
As far as I can tell, we always use `npm install` or `npm ci` from the
project root to install dependencies for building the app, so we don't
actually ever use these. All the dependencies are already in the main
package.json.