Guarantee ordering of data seeding
The GitHub actions are currently frequently failing because of failing
migrations. This appears to be due to out of order inserts to the
instruments table, which references itself via the parent field. When
inserting an instrument, its parent must exist to satisfy the foreign
key constraints. If the parent does not exist, this will fail.

Since none of the `em.upsert` calls were `await`ed, none of the changes
were persisted until we called flush, at which point they could be
applied in any order, causing intermittent foreign key constraint
violations.