Update types in cal.php and fix login requiring cookies
Many of the types in `data/cal.php/+server.ts` were incomplete or
outdated (due to changes since mikro-orm was added). This fixes all the
type errors in that file.

Additionally, the tests, and all my manual testing were overlooking how
the authentication needs to work for the iCal files. Notably, the query
parameters determine which user is logged in, not the session. This
allows e.g. Google calendar to fetch the link and update the calendar
entries periodically. With these changes, testAuthLink is used to get
the user id that is used to generate the session, rather than relying on
the session information retrieved from the cookie we set.