Mount images from block storage to sveltekit to enable profile pictures
SvelteKit needs to be able to access the images directory to serve
profile pictures (and also to handle uploads). Caddy cannot serve these
images as they must only be accessed by authenticated users.

Additionally, this changes the modified date for images from a
date string to an ISO string as this includes time information and
never contains a space, both are desirable given this is being used for
a query parameter in a URL (to hint to browsers when a cached version
of the image can and cannot be used).